Added get notifications

+++ src/main/kotlin/com/vf/business/business/dao/repo/ConversationCorrespondentRepository.kt
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface ConversationCorrespondentRepository: CrudRepository<ConversationCorrespondent, Int> {
+interface ConversationCorrespondentRepository: CrudRepository<ConversationCorrespondent, Long> {
 }
