+++ src/main/kotlin/com/vf/business/business/dao/repo/ConversationMessageRepository.kt
@@ -8,7 +8,7 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface ConversationMessageRepository: CrudRepository<ConversationMessage, Int> {
+interface ConversationMessageRepository: CrudRepository<ConversationMessage, Long> {
 
     fun findByConversationOrderByCreatedAtDesc(conversation: Conversation, pageReq: PageRequest): Page<ConversationMessage>
 }
